Gui: split out some basic coin3d related definition
