+++ src/Gui/ViewProvider.h
@@ -31,12 +31,12 @@
 #include <bitset>
 #include <QIcon>
 #include <boost/signals2.hpp>
-#include <boost/intrusive_ptr.hpp>
 
 #include <App/TransactionalObject.h>
 #include <App/Material.h>
 #include <Base/Vector3D.h>
 #include <Base/BoundBox.h>
+#include "InventorBase.h"
 
 class SbVec2s;
 class SbVec3f;
@@ -83,32 +83,6 @@ enum ViewStatus {
     TouchDocument = 4,
 };
 
-
-/** Convenience smart pointer to wrap coin node. 
- *
- * It is basically boost::intrusive plus implicit pointer conversion to save the
- * trouble of typing get() all the time.
- */
-template<class T>
-class CoinPtr: public boost::intrusive_ptr<T> {
-public:
-    // Too bad, VC2013 does not support constructor inheritance
-    //using boost::intrusive_ptr<T>::intrusive_ptr;
-    typedef boost::intrusive_ptr<T> inherited;
-    CoinPtr() {}
-    CoinPtr(T *p, bool add_ref=true):inherited(p,add_ref){}
-    template<class Y> CoinPtr(CoinPtr<Y> const &r):inherited(r){}
-
-    operator T *() const {
-        return this->get();
-    }
-};
-
-/** Helper function to deal with bug in SoNode::removeAllChildren()
- *
- * @sa https://bitbucket.org/Coin3D/coin/pull-requests/119/fix-sochildlist-auditing/diff
- */
-void GuiExport coinRemoveAllChildren(SoGroup *node);
 
 /** General interface for all visual stuff in FreeCAD
   * This class is used to generate and handle all around
